Fix front door flood lighting in HA 2021.12
Fix front_door_end_flood for HA 2021.12. Apparenty, HA or the Hue
integration changed how current RGB state of a light group is
returned when only part of the group is turned on. (Used to be the
value of the "on" lights. Now it seems to be some kind of average.)
So this commit checks RGB values of individual lights instead of
the group when deciding to turn off flood lighting.